Add first version of upload() function
It’s not called yet, and I haven’t tested it, but I hope it will work.
It’s loosely based on the Python chunked upload sample code in the API
documentation [1], but without special handling of the first chunk. We
don’t know the file size in advance (the source files are gzipped, and I
plan to decompress them on-the-fly), so we send a fake file size to the
API; as far as I can tell from looking at ApiUpload.php, the API doesn’t
check that it’s the same for each upload in a chunk, the only important
thing is that the last chunk is marked by the total bytes being equal to
the file size in that request.

read_chunks() has a special case for ensuring that even if the input
file happens to fit exactly into the last chunk, it sends one more empty
chunk, which should cause upload() to send the final request where the
file size matches the total bytes. However, I haven’t tested it, and
given that upload() picks a chunk size of one mebibyte, this case is
exceedingly unlikely to happen.

[1]: https://www.mediawiki.org/wiki/API:Upload#Sample_code_3