Add some categories to the uploaded files
We decided that categories should not completely be added by the {{We
Are Beautiful model}} template, but rather be specified on the file
itself, so that more accurate categories can be set on each file by
individual editors if applicable. For the initial upload, guess a first
selection of templates.